Move Overwatch runner to its own module.
